Normalize ModOperator struct to be consistent with other operators
The ModOperator was a bit of an outcast before, in that the name was
ModOperation while other arithmetic operators were XOperator. ModOperation
also had "A" and "B" values while everything else had "Left" and "Right".

This makes the ModOperator more consistent with other types, and also updates
it to properly handle precedence properly, as it's now doesn't require the
special cases that it previously did.