Refactor CALL IR opcode to store list of arguments
Refactor IR for CALL to keep a slice of argument registers, rather
than moving them into a register in separate MOV instructions.

This should make it easier to adopt different calling conventions in the
code generation, instead of assuming them in the IR generation. It also
lays the framework for tail call optimization (which requires a different
convention for where to put the parameters than a normal CALL instruction,
since it re-uses the existing stack space.)