Implemented rules for variable shadowing
This adds rules for mutable and immutable variable shadowing.

let statements may be shadowed with other statements of the same name
(which are distinct variables, that just happen to have the same name.)

mut statements, on the other hand, can never be shadowed, even by variables
in another scope. Since assigning to shadoweded variables often causes
confusion (and bugs) in imperative languages, you're forced to choose
a new name for mutables, not reuse an existing one.