Add support for generic enumerations to AST
Added the ability for the AST to parse parameterized enumerations such
as `data Foo x = A x | B x | C`.

Concrete types based on Foo may then be created by specifying the x at
declaration time, such as `func foo() (Foo int)` to declare a function
foo which returns a "Foo int" type, or `let x Foo string = B
"something"` to create variable of type `Foo string` (and initialize
it to the enumerated option `B` with the parameter of `"something"`.)

Code generation for the parsed AST is not yet implemented.