Add a new test program and fixed (some) of the bugs it exposed
Add a new test program, which runs a meaningless algorithm, but exposed
some bugs and fixed those bugs.

- Let statements that used a variable in the initializer with the same
  name were incorrectly interpreting it as a reference to the new variable
- Not all operators were acting as token delimiters in the tokenizer
- The new sample program was tripping over the (known) issue of the code
  generation using a constant stack size for each function, rather than
  an appropriate one for the function being generated. This doesn't fix
  the bug, it just increases the size from 32 to 40 as a temporary hack.

In addition to this, the (unused) helloworld2 sample program is removed,
and a "-debug" parameter is added to the l command, which prints more
information about what it's doing and doesn't delete temporary files.