Added basic arrays
Added the ability to create arrays or array literals with
syntax similar to Go's. Currently, only constant length
arrays are supported, and they aren't well tested. In particular
there's not as many checks as there should be for things like
indexing out of bounds, so if you do something illegal it's likely
to compile, but crash.