Added preliminary support for reference parameters.
procs may now have parameters defined as reference parameters by prefixing
them with the "mutable" keyword. This is primarily driven by the fact that
a Read() syscall needs to read into one of the parameters.

There should be more restrictions on the mutable keyword/reference parameter
than there currently are. For instance, only mutable variables should be
passed by reference, pure functions shouldn't allow mutable parameters, etc.

For now, this is only intended to do the minimum required to enable the
implementation of an OS specific Read to enable file I/O. (On this front,
arrays/slices or malloc will also be required for it to be able to read into
an arbitrarily sized structure.)