Move bad programs test from compiler to AST package
The bad programs should have never made it to the code generation (or
even IR generation) stages, so this speeds them up by moving it to the
parser, where it doesn't need to compete with the (slow) valid program
tests.

Also removed the "print" hack from the AST side of things, since it's
no longer used.