Add basic support for slices
Slices differ from arrays in that the size of them isn't known, and
the size gets carried around before the data (and passed in to functions along
with the data.)

This differs from most languages, where slices contain a pointer to an array,
rather than simply having a prefix before the array.

The primary reason for this difference is because there are no pointers in
yet (or at least, only very preliminary support), not because this is a better
design (in fact, it's far more inefficient). It will likely get updated in the
future, but for now slices exist only to add the basics of what will be required
in order to add a Read() syscall.