Added the ability to do basic file I/O and support for OSX and Linux
Rewrote the printing in terms of wrappers around I/O syscalls. The
open call is based around Plan9/Go, rather than POSIX, because it's
easier impement the interface for that in terms of POSIX than it is
to wrap the POSIX interface around Plan9's syscalls, and the interface
is also slightly easier to use as an API.

This adds support for the following procedures as built in:

	proc Open(filename string) (int)
	proc Create(filename string) (int)
	proc Close(fd int) (int)
	proc Read(fd int, []byte buf) (int)
	proc Write(fd int, []byte buf) (int)

They'll likely remain built in until namespaces are implemented,
at which point they'll move into a file (or similar) package.

Print is expanded to include PrintInt, PrintString, and PrintByteSlice.

OSX and Linux support is added, since they all use the same system call
convention on amd64, and the only difference is the syscall numbers and
constants for Open/Create. Even though they aren't used, they're the
only operating systems supported by travis, so this enables the possibility
of having some basic continuous integration for the tests.