Added len method for []string
Added the len method to check the slice length of a slice of strings.

This should be more generic across all slice types (the assembly doesn't
change), but since there's neither multiple dispatch nor generic functions
implemented, some type had to go in the function signature (or there'd have
to be hacks added), and strings won since (eventually) proc main will take
a []string as a parameter, meaning that len([]string) needs to work in
order to provide access to command line arguments.