Added TempValue type to IR to differentiate variables from intermediary calculations
Previously, everything was stored as a LocalValue type in the IR. This meant that everything
had to be saved to a memory location, since there was no way for the code generation to tell
the difference between a variable in memory and and intermediate calculations that's only
used once.

The TempValue represents values that are ephemeral, such as the x+1 in `let x = x + 1` and
usually only ever exist in a (physical hardware) register when doing code generation. This
should save a considerable amount of memory on the stack.