Fixed bug where strings couldn't be indexed in an array
Since the size in bytes of a string isn't really meaningful,
the value of "0" was previously used. Unfortunately, this meant
that when calculating the offset into an array for a string, the
offset of SizeOfType*index was always calculated as zero.

The size of a string is now internally stored as "8" (the size
of a pointer), rather than "0", to faciliate offset calculations.