Added the ability to pass arguments to proc main.
main can now have the signature "proc main(args []string) ()" instead of
"proc main () ()" and receives the command line arguments as an argument.

Added a simple "echo" program as a test case for passing arguments to the
main function, and fixed the bugs that it uncovered. This is the first
step to being able to do something interesting.