Reserve a more appropriate stack size for functions
Previously, the amount of stack space that was reserved for
a function definition was a hardcoded constant for testing purposes.

This updates the code to be (very slightly) more intelligent, and instead
reserve a (still non-optimal) amount of stack space based on the needs
of the function. It reserves space for any local variables, space for
function parameters, and space for function parameters that it passes
to function calls.

The latter shouldn't be required, but at the moment seems to be required
to prevent corrupting the stack. This should be revisited later.