Added higher level IR representation
Add higher level IR representation ("hlir") than the previously
existing one. While the previous IR was essentially amd64 asm with
an infinite number of (typed) registers, the new hlir is
a jump-less IR based on it, with jumps replaced by IF, LOOP.
and JumpTable instructions, so that control flow only moves forward.

This serves 2 purposes:
1. register usage can now be analyzed to eliminate dead code more easily,
   since we know that the IR will never jump backwards to a previous instruction
   which used a register. (outside of a loop, inside which flow still only
   moves forwards)
2. It should be easier to compile this new IR to WebAssembly.

Neither of those benefits are realized yet, (in fact, the original IR hasn't
even been updated yet to be generated from hlir instead of the ast) but it
should make it easier to realize these benefits them in the future.