Added WASM Intermediate representation
Add a WASM representation generated from the HLIR added in the previous commit. The
WASM representation currently passes all the non-file related test cases passed by
the native tests, but are not yet implemented by the "l" compiler command. In order
to convert the module to a useable binary, the Module returned by wasm.Parse's stringer
method can be called. This will print the Module in WAST format, which can then be
compiled by a WAST compiler such as https://cdn.rawgit.com/WebAssembly/wabt/fb986fbd/demo/wat2wasm/

Serialization directly to WASM binary format is not yet implemented.

The built-in functions PrintInt, PrintString, PrintByteSlice, and len are automatically
imported, but must be written in javascript for the WebAssembly instance using in it
a way that makes sense.

PrintString takes an address which consists of a uint64 followed by that many bytes in
utf8 format.

PrintInt takes a int32. (i64s are wrapped to ensure they work in a web browser context.)

PrintByteSlice takes an int array length and an int address of the start.

len takes arguments of the same format as PrintByteSlice, but should only return the first.