Generate jump based machine level IR from hlir package
The hlir package is now used to generate both the machine code, and
the wasm generation. This provides a centralized place to implement
optimizations before the code paths diverge.

The package "ir" is also renamed "mlir", because it no longer makes
sense to simply call it "ir" with multiple different intermediate
representations.