Convert strings and byte slices to same internal representation
Previously, strings were stored as n followed by n bytes, while
byte slices were stored as n followed by a pointer to n bytes. This
harmonizes them to have the same representation (the latter) so that
Write() can be called with either a string or a byte slice.

WASM tests are temporarily broken, because it hasn't been updated
to reflect changes in the HLIR which were required for this change,
or the fact strings are now passed to function calls as len, strbase
and not as a pointer to a memory location which has len, strbase
located at it.