Convert PrintByteSlice to be a native function
PrintByteSlice is now implemented in the language itself (by
calling Write(1, ..)), making it the first part of the standard
library bootstrapping process.

The "standard library" (containing only PrintByteSlice) is currently
implemented as a string in Go from the compiler, and should eventually
become hosted on the filesystem, but for now this allows the compiler
to work with "get get" and not require any environment setup.