Convert compiler tests from Go strings to real files
Converted the "sampleprograms" from Go strings to normal files
that are read from the test suite. This makes it easier to add
new tests and will eventually allow the "sampleprograms" package
to be removed, but for now it remains because packages other than
codegen haven't been updated to use it yet.