Switch from Go tools to just invoking `go build`
The Go linker is not feasible to use for this language because it
makes too many assumptions about the Go runtime, and the stack splitting
in particular makes it impossible to use to assemble any non-trivial
programs.

This changes the code to instead create a Go program which calls the
generated assembly, after converting the OS args to our internal
format.

This results in (much) larger binaries because of all the unnecessary
overhead that Go adds, but since this is only intended to be a bootstrap
compiler in the long run it shouldn't matter, once we get to the point
where we're directly emitting our own code.